evalv6 + training consistency: use bf16 everywhere.

+++ rlhf/grpo/my_grpo.py
@@ -188,7 +188,9 @@ if __name__ == "__main__":
         )
 
     policy = AutoModelForCausalLM.from_pretrained(
-        model_args.model_name_or_path, trust_remote_code=model_args.trust_remote_code
+        model_args.model_name_or_path,
+        trust_remote_code=model_args.trust_remote_code,
+        torch_dtype=torch.bfloat16,
     )
 
     ################
